+++ poetry.toml
@@ -1,0 +1,3 @@
+[virtualenvs]
+create = true
+in-project = true
