+++ pyproject.toml
@@ -1,0 +1,28 @@
+[tool.poetry]
+name = "vagrant"
+version = "0.1.0"
+description = "x"
+authors = ["x"]
+license = "x"
+
+[tool.poetry.dependencies]
+python = "^3.7"
+certifi = "2020.6.20"
+chardet = "3.0.4"
+click = "7.1.2"
+Flask = "1.1.1"
+idna = "2.10"
+itsdangerous = "1.1.0"
+Jinja2 = "2.11.2"
+MarkupSafe = "1.1.1"
+python-dotenv = "0.12.0"
+requests = "2.24.0"
+urllib3 = "1.25.10"
+Werkzeug = "1.0.1"
+python-dateutil = "2.8.1"
+
+[tool.poetry.dev-dependencies]
+
+[build-system]
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
