+++ todo_app/app.py
@@ -1,12 +1,12 @@
 from flask import Flask, render_template, request, redirect, url_for
-import session_items
-from view_model import ViewModel
+import todo_app.data.session_items as session_items
+from todo_app.view_model import ViewModel
 from datetime import datetime, timedelta
 
 
 def create_app():
     app = Flask(__name__)
-    app.config.from_object("flask_config.Config")
+    app.config.from_object("todo_app.flask_config.Config")
 
     @app.route("/")
     def index():
