+++ todo_app/templates/index.html
@@ -3,21 +3,20 @@
 
 {% block content %}
 <div class="jumbotron"
-  style="background:url(https://png.pngtree.com/thumb_back/fh260/background/20190223/ourmid/pngtree-creative-color-splashing-wind-color-paint-background-backgroundpigmentsplashwatercolor-windcoatingmulticolored-backgroundbanner-image_68593.jpg) ; background-size: 100% 100%;">
-  <h1 class="display-1 text-center" style="color:rgb(9, 168, 226)">&#128151; Lindsey J Raven &#128151;</h1>
+  style="background:url(https://png.pngtree.com/thumb_back/fh260/background/20200714/pngtree-modern-double-color-futuristic-neon-background-image_351866.jpg) ; background-size: 100% 100%;">
+  <h1 class="display-1 text-center" style="color:rgb(245, 247, 248)">&#128151; Lindsey J Raven &#128151;</h1>
 </div>
 
-
-<body style="background:url(https://wallpaperaccess.com/full/545053.jpg) ; background-size: 100% 100%;"></body>
+<body style="background:url(https://coolbackgrounds.io/images/backgrounds/index/sea-edge-79ab30e2.png) ; background-size: 100% 100%;"></body>
 
 
 <div class="col-md-12 text-center"
-  style="background:url(https://previews.123rf.com/images/hakkiarslan/hakkiarslan1601/hakkiarslan160100032/51690961-abstract-color-splash-background.jpg) ; background-size: 100% 100%;">
+  style="background:url(https://png.pngtree.com/thumb_back/fh260/background/20200714/pngtree-modern-double-color-futuristic-neon-background-image_351866.jpg) ; background-size: 100% 100%;">
   <span style='font-size:30px;'></span>
   <form method="POST" action="/add_item">
     </br>
     </br>
-    <label style="color:#0373bd" for="name">Enter your to-do:</label>
+    <label style="color:#fcfcfc" for="name">Enter your to-do:</label>
     <input class="text-info" type="text" name="name">
     <input class="text-info" type="submit" class="btn btn-dark btn-md" value="Submit">
   </form>
