+++ todo_app/view_model.py
